remove specific hooks + add hook customization
