+++ vcr/vcr.go
@@ -1,9 +1,6 @@
 package vcr
 
 import (
-	"encoding/base64"
-	"encoding/json"
-	"encoding/xml"
 	"net/http"
 	"path/filepath"
 	"regexp"
@@ -14,17 +11,6 @@ import (
 	"gopkg.in/dnaeon/go-vcr.v4/pkg/cassette"
 	"gopkg.in/dnaeon/go-vcr.v4/pkg/recorder"
 )
-
-const (
-	windowsDockerEngine     = "//./pipe/docker_engine"
-	unixDockerEngine        = "/var/run/docker.sock"
-	escapedUnixDockerEngine = "%2Fvar%2Frun%2Fdocker.sock"
-)
-
-// QueryMatcherIgnore contains the list of query value that should be ignored when matching requests with cassettes
-var QueryMatcherIgnore = []string{
-	"organization_id",
-}
 
 // getTestFilePath returns a valid filename path based on the go test name and suffix. (Take care of non fs friendly char)
 func getTestFilePath(t *testing.T, pkgFolder string, suffix string) string {
@@ -53,7 +39,6 @@ func cassetteRequestFilter(i *cassette.Interaction) error {
 	delete(i.Request.Headers, "Authorization")
 
 	// URL substitutions
-	// 	Make URLs generic
 	i.Request.URL = regexp.MustCompile(`(.+)organization_id=[0-9a-f-]{36}(.+)`).
 		ReplaceAllString(i.Request.URL,
 			"${1}organization_id=11111111-1111-1111-1111-111111111111${2}")
@@ -65,15 +50,6 @@ func cassetteRequestFilter(i *cassette.Interaction) error {
 		ReplaceAllString(
 			i.Request.URL,
 			"${1}SCWXXXXXXXXXXXXXXXXX${2}")
-	//	Buildpacks
-	i.Request.URL = regexp.MustCompile(`pack\.local%2Fbuilder%2F[0-9a-f]{20}`).
-		ReplaceAllString(i.Request.URL, "pack.local%2Fbuilder%2F11111111111111111111")
-	i.Request.URL = regexp.MustCompile(`pack\.local/builder/[0-9a-f]{20}`).
-		ReplaceAllString(i.Request.URL, "pack.local/builder/11111111111111111111")
-
-	// Body substitutions
-	i.Request.Body = regexp.MustCompile(`pack\.local/builder/[0-9a-f]{20}`).
-		ReplaceAllString(i.Request.Body, "pack.local/builder/11111111111111111111")
 
 	return nil
 }
@@ -87,46 +63,12 @@ func cassetteResponseFilter(i *cassette.Interaction) error {
 	i.Response.Body = regexp.MustCompile(`"secret_key":"[0-9a-f-]{36}"`).
 		ReplaceAllString(i.Response.Body, `"secret_key":"11111111-1111-1111-1111-111111111111"`)
 
-	i.Response.Body = regexp.MustCompile(`pack\.local/builder/[0-9a-f]{20}`).
-		ReplaceAllString(i.Response.Body, "pack.local/builder/11111111111111111111")
-
 	return nil
 }
 
-// This hook is needed for container deploy tests on the CLI
-func unescapeDockerURL(i *cassette.Interaction) error {
-	i.Request.URL = regexp.MustCompile(`http://`+escapedUnixDockerEngine+`(.+)?`).
-		ReplaceAllString(
-			i.Request.URL,
-			"http://"+unixDockerEngine+"${1}")
-
-	return nil
-}
-
-func s3Encoder(i *cassette.Interaction) error {
-	if strings.HasSuffix(i.Request.Host, "scw.cloud") {
-		if i.Request.Body != "" && i.Request.Headers.Get("Content-Type") == "application/octet-stream" {
-			requestBody := []byte(i.Request.Body)
-			if !json.Valid(requestBody) {
-				err := xml.Unmarshal(requestBody, new(any))
-				if err != nil {
-					i.Request.Body = base64.StdEncoding.EncodeToString(requestBody)
-				}
-			}
-		}
-
-		if i.Response.Body != "" && i.Response.Headers.Get("Content-Type") == "binary/octet-stream" {
-			responseBody := []byte(i.Response.Body)
-			if !json.Valid(responseBody) {
-				err := xml.Unmarshal(responseBody, new(any))
-				if err != nil {
-					i.Response.Body = base64.StdEncoding.EncodeToString(responseBody)
-				}
-			}
-		}
-	}
-
-	return nil
+type AdditionalHook struct {
+	HookFunc recorder.HookFunc
+	Kind     recorder.HookKind
 }
 
 // NewHTTPRecorder creates a new httpClient that records all HTTP requests in a cassette.
@@ -135,7 +77,7 @@ func s3Encoder(i *cassette.Interaction) error {
 //
 // It is important to add a `defer cleanup()` so the given cassette files are correctly
 // closed and saved after the requests.
-func NewHTTPRecorder(t *testing.T, pkgFolder string, update bool, realTransport ...http.RoundTripper) (*recorder.Recorder, error) {
+func NewHTTPRecorder(t *testing.T, pkgFolder string, update bool, realTransport http.RoundTripper, additionalHooks ...AdditionalHook) (*recorder.Recorder, error) {
 	t.Helper()
 
 	recorderOptions := []recorder.Option{
@@ -145,12 +87,6 @@ func NewHTTPRecorder(t *testing.T, pkgFolder string, update bool, realTransport 
 		recorder.WithHook(cassetteRequestFilter, recorder.BeforeSaveHook),
 		// Remove secrets and unnecessary information from the response
 		recorder.WithHook(cassetteResponseFilter, recorder.BeforeSaveHook),
-		// Starting with v3, go-vcr now calls net/url.Parse to build the interaction which results in an error for escaped
-		// Docker URLs on Test_Deploy (container), so we need to unescape these paths after capture
-		recorder.WithHook(unescapeDockerURL, recorder.AfterCaptureHook),
-		// Serialization changed on v4, we now need to encode binary payloads as base64
-		recorder.WithHook(s3Encoder, recorder.AfterCaptureHook),
-
 		// Add custom matcher for requests and cassettes
 		recorder.WithMatcher(CassetteMatcher),
 	}
@@ -162,9 +98,14 @@ func NewHTTPRecorder(t *testing.T, pkgFolder string, update bool, realTransport 
 	}
 	recorderOptions = append(recorderOptions, recorder.WithMode(recorderMode))
 
-	// Set up custom transport for the CLI
-	if len(realTransport) > 0 {
-		recorderOptions = append(recorderOptions, recorder.WithRealTransport(realTransport[0]))
+	// Set up custom transport (for the CLI only)
+	if realTransport != nil {
+		recorderOptions = append(recorderOptions, recorder.WithRealTransport(realTransport))
+	}
+
+	// Set up specific hooks for each DevTool
+	for _, hook := range additionalHooks {
+		recorderOptions = append(recorderOptions, recorder.WithHook(hook.HookFunc, hook.Kind))
 	}
 
 	// Setup recorder and scw client
